LayoutTests/http/tests/security/contentSecurityPolicy/1.1/scriptnonce-badnonce.html

<!DOCTYPE html>
<html>
    <head>
        <meta http-equiv="X-WebKit-CSP" content="script-nonce this is a bad nonce;">
        <script>
        if (window.layoutTestController)
            layoutTestController.dumpAsText();
        </script>
        <script>
            alert('PASS (1/1)');
        </script>
    </head>
    <body>
        <p>
            This tests the effect of an invalid script-nonce value. It passes if
            a console warning is visible, and the alert() is executed.
        </p>
    </body>
</html>
