LayoutTests/http/tests/security/contentSecurityPolicy/1.1/scriptnonce-badnonce-expected.txt

CONSOLE MESSAGE: Ignoring invalid Content Security Policy script nonce: 'this is a bad nonce'.

ALERT: PASS (1/1)
This tests the effect of an invalid script-nonce value. It passes if a console warning is visible, and the alert() is executed.
